requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Deep Learning & Embeddings
sentence-transformers>=2.2.0
torch>=2.0.0

# FAISS
faiss-cpu>=1.7.4

# Gradient Boosting
xgboost>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
joblib>=1.3.0
tqdm>=4.65.0
gitpython>=3.1.0